minor tweaks to model and node 2 not forwarding complex events to node 3 anymore
The latter is for the experiments only, since no network communication will happen in the real experiment

+++ ns-3-dev/src/dcep/examples/MANETSimulation.cc
@@ -61,7 +61,7 @@ int main(int argc, char** argv) {
     uint32_t numMobile = 7;
     uint32_t allNodes = numMobile+numStationary;
     uint64_t stateSize = 100;
-    uint32_t eventInterval = 1000000;  // Interval in nanoseconds
+    uint64_t eventInterval = 1000000;  // Interval in nanoseconds
     
     std::string format ("OMNet++");
     std::string experiment ("dcep-performance-test"); //the current study
